Esp32 initial patches (#1)
* decoder: widen glyph metrics and width APIs to 16 bits

The font metrics and the character width returns were all uint8_t, which
caps usable glyphs at 255 px. Fonts of 175-200 px are within a factor of
two of that ceiling, and a wide glyph in a large face silently truncates.

Widen the mf_font_s metrics (width, height, min/max_x_advance, baseline_x,
baseline_y, line_height) and the character_width / render_character
function pointers to 16 bits, along with every implementation
(mf_rlefont, mf_bwfont, mf_scaledfont) and the public mf_render_character,
mf_character_width, mf_character_whitespace and mf_character_callback_t.

mf_compute_kerning returns int16_t for the same reason: its clamp is a
percentage of the glyph width, so a 300 px glyph produces an adjustment
that does not fit in int8_t. The kerning edge tracking moves to uint16_t
with a UINT16_MAX sentinel.

This is an API break for callers that declare their own character
callbacks; the on-disk font format is unchanged by this commit.


* rlefont: store glyph width as 16 bits (format version 5)

With the metrics widened, the per-glyph width in the encoded data was
still a single byte and remained the hard 255 px cap.

Store it as two little-endian bytes. The decoder reads it byte-wise
rather than through pgm_read_word: the glyph data is a packed uint8_t
array, so a 16-bit load lands on an arbitrary offset and is both an
unaligned access and a strict-aliasing violation.

Two accounting fixes go with it:

  - get_glyph_size() still reserved one byte per glyph. It feeds
    compute_char_ranges() against the 65536 limit imposed by the
    uint16_t glyph offsets, so underestimating by a byte per glyph can
    let a range overflow its offset table and produce a font that
    decodes to garbage.

  - The high byte was written as `width & 0xFF >> 8`. Shift binds
    tighter than bitwise and, so `0xFF >> 8` is 0 and the high byte was
    always zero.

The format is not backwards compatible, so bump RLEFONT_FORMAT_VERSION
to 5 and drop MF_RLEFONT_VERSION_4_SUPPORTED. Version 4 .c files now
fail the #error guard at compile time instead of mis-decoding; re-run
the encoder to regenerate them.

The bwfont format is unchanged and still caps glyph width at 255.


* encoder: do not build the unit tests by default

`make` ran run_unittests first, which needs cxxtestgen to generate
unittests.cc. That tool is not commonly installed and is not needed to
encode a font, so a plain `make` failed on most machines.

Leave the rules in place and build them with `make run_unittests`.


* encoder: add a -v/--verbose switch for progress output

Importing and optimizing a large font is slow enough to look hung.
optimize() already threaded a verbose flag through every pass but
hardcoded it to false, and there was no way to see which glyph a slow
freetype import was working on.

Add a global flag set by -v / --verbose before command dispatch, use it
for the optimizer passes, and print the codepoint being imported.
Default stays quiet so scripted font builds are unaffected.


* examples: widen render_bmp character callback to uint16_t

Follows the mf_character_callback_t signature change.


* Add an ESP-IDF component manifest

Lets the repository be used directly as an ESP-IDF component (as a
submodule under components/) with no wrapper on the project side.

Builds the decoder only; the encoder is a host tool driven from
fonts/Makefile. The generated fonts.h is project data rather than
library data, so its location comes from MCUFONT_FONT_DIR and falls back
to this repository's own fonts/ directory.

Also ignore the host encoder and example binaries.


---------

+++ decoder/mf_bwfont.c
@@ -102,9 +102,9 @@ static uint8_t render_char(const struct mf_bwfont_char_range_s *r,
     return get_width(r, index);
 }
 
-uint8_t mf_bwfont_render_character(const struct mf_font_s *font,
-                                   int16_t x0, int16_t y0,
-                                   uint16_t character,
+uint16_t mf_bwfont_render_character(const struct mf_font_s *font,
+                                    int16_t x0, int16_t y0,
+                                    uint16_t character,
                                    mf_pixel_callback_t callback,
                                    void *state)
 {
@@ -119,8 +119,8 @@ uint8_t mf_bwfont_render_character(const struct mf_font_s *font,
     return render_char(range, x0, y0, index, callback, state);
 }
 
-uint8_t mf_bwfont_character_width(const struct mf_font_s *font,
-                                  uint16_t character)
+uint16_t mf_bwfont_character_width(const struct mf_font_s *font,
+                                   uint16_t character)
 {
     const struct mf_bwfont_s *bwfont = (const struct mf_bwfont_s*)font;
     const struct mf_bwfont_char_range_s *range;
