+++ decoder/mf_bwfont.h
@@ -64,14 +64,14 @@ struct mf_bwfont_s
 
 #ifdef MF_BWFONT_INTERNALS
 /* Internal functions, don't use these directly. */
-MF_EXTERN uint8_t mf_bwfont_render_character(const struct mf_font_s *font,
+MF_EXTERN uint16_t mf_bwfont_render_character(const struct mf_font_s *font,
                                              int16_t x0, int16_t y0,
                                              mf_char character,
                                              mf_pixel_callback_t callback,
                                              void *state);
 
-MF_EXTERN uint8_t mf_bwfont_character_width(const struct mf_font_s *font,
-                                            mf_char character);
+MF_EXTERN uint16_t mf_bwfont_character_width(const struct mf_font_s *font,
+                                             mf_char character);
 #endif
 
 #endif
