+++ decoder/mf_kerning.h
@@ -18,7 +18,7 @@
  * Returns the offset to add to the x position for c2.
  */
 #if MF_USE_KERNING
-MF_EXTERN int8_t mf_compute_kerning(const struct mf_font_s *font,
+MF_EXTERN int16_t mf_compute_kerning(const struct mf_font_s *font,
                                     mf_char c1, mf_char c2);
 #else
 #define mf_compute_kerning(f,c1,c2) (0)
